tradebot/requirements.txt
# ── Broker & Market Data ──────────────────────────────────────
kiteconnect==4.2.0
smartapi-python==1.3.5
nsepython==2.7
yfinance==0.2.38

# ── Data Engineering ──────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b
scipy==1.13.0
pyti==0.2.2

# ── Machine Learning ──────────────────────────────────────────
lightgbm==4.3.0
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.44.1
prophet==1.1.5
h2o==3.46.0.1

# ── NLP & Sentiment ───────────────────────────────────────────
transformers==4.40.0
torch==2.2.2
openai==1.23.6
openai-whisper==20231117
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
deep-translator==1.11.4

# ── Orchestration ─────────────────────────────────────────────
apache-airflow==2.8.4
pendulum==3.0.0
celery==5.3.6
redis==5.0.4
schedule==1.2.1

# ── Storage ───────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# ── API & Realtime ────────────────────────────────────────────
fastapi==0.110.2
uvicorn==0.29.0
websockets==12.0
pydantic==2.7.1
python-multipart==0.0.9

# ── Monitoring ────────────────────────────────────────────────
mlflow==2.12.2
evidently==0.4.30

# ── Notifications ─────────────────────────────────────────────
firebase-admin==6.5.0

# ── Utils ─────────────────────────────────────────────────────
joblib==1.4.0
python-dotenv==1.0.1
pyyaml==6.0.1
httpx==0.27.0
